django_ai_service/requirements.txt
django
djangorestframework
djongo
pymongo
pika
requests
numpy
pandas
scikit-learn
openai-whisper
opencv-python
